feat: modify `Image` entity
`sha1` to `sha256`, `size` type to Long

+++ app/src/main/kotlin/io/sakurasou/controller/vo/UserVO.kt
@@ -17,7 +17,7 @@ data class UserVO(
     val isBanned: Boolean,
     val createTime: LocalDateTime,
     val imageCount: Long,
-    val totalImageSize: Double
+    val totalImageSize: Long
 )
 
 @Serializable
@@ -28,5 +28,5 @@ data class UserPageVO(
     val isBanned: Boolean,
     val createTime: LocalDateTime,
     val imageCount: Long,
-    val totalImageSize: Double
+    val totalImageSize: Long
 )
